fix(core): split date normalization out and declare its parser
Three follow-ons to the #252 fix, each found by a check rather than by
reading:

* CI proved the parser was never there. `python-dateutil` was not a
  declared dependency and is absent from every `requirements/ci-*.txt`,
  so all four Test jobs ran with it missing — the dateutil branch of
  `normalize_date_to_iso` was dead code in CI, and any install that
  resolved without it silently kept dates it could not read (the LoCoMo
  shape the docstring itself cites). Declared, locked, and re-exported
  into the 9 hash-pinned requirement sets.

* The guarantee now holds on the degraded path too. When the parse
  produces nothing and the string STATES a zone, the value is refused
  instead of salvaged: salvaging returns a naive date, and a naive date
  is read as UTC by exactly the consumers this issue is about. A string
  with no zone still salvages its date, as before.

* `normalize_date_to_iso` moves to `core/temporal_normalize.py`.
  Storage normalization ("which instant does this string denote?") and
  retrieval scoring ("how close is this date to the query?") have
  different reasons to change — and the fix had pushed temporal.py to
  302 lines, past this repo's 300-line cap.
